Movved redemption function from utils to GP ts file
