+++ src/utils.ts
@@ -19,42 +19,6 @@ export function concat(a: ByteArray, b: ByteArray): ByteArray {
   return out as ByteArray;
 }
 
-export function updateRedemption(
-  beneficiary: Address,
-  avatar: Address,
-  amount: BigInt,
-  proposalId: Bytes,
-  rewardType: ByteArray,
-  rewardString: String
-): void {
-  let accountId = crypto.keccak256(concat(beneficiary, avatar));
-
-  let rewardId = crypto.keccak256(concat(rewardType, amount as ByteArray));
-
-  let uniqueId = crypto
-    .keccak256(concat(proposalId, concat(accountId, rewardId)))
-    .toHex();
-
-  let redemption = store.get("Redemption", uniqueId) as Redemption;
-  if (redemption == null) {
-    redemption = new Redemption();
-    redemption.redeemer = beneficiary;
-    redemption.proposalId = proposalId.toHex();
-    redemption.rewardId = rewardId.toHex();
-    store.set("Redemption", uniqueId, redemption);
-  }
-
-  let reward = store.get("Reward", rewardId.toHex()) as Reward;
-  if (reward == null) {
-    reward = new Reward();
-    reward.id = rewardId.toHex();
-    reward.type = rewardString.toString();
-    reward.amount = amount;
-
-    store.set("Reward", rewardId.toHex(), reward);
-  }
-}
-
 export const zero256 =
   "0x0000000000000000000000000000000000000000000000000000000000000000";
